Save abandoned attempt at converting to TypeScript
This gets really hard because I want to use TypeScript w/ generators.

To use it with generators, I need to set a compile target of ES6, which means
I also need to post-process with babel.

That's all fine and good, but then I can't run tests in an obvious way, because
I'd need to make my own typescript+babel/register script to force node to go
through a compile step before execution of the modules.

This is hard :<